workshops/notion-agents/requirements.txt
mcp>=1.7.1
httpx>=0.27.2
uvicorn[standard]>=0.30.6
starlette>=0.41.0
python-dotenv>=1.0.1
